Add files via upload
 Here,the weather dataset is a time series data set with per-hour infromation about the weather conditions at a particular location .It records tempreture,dew point tempreture,relative humidity ,wind speed ,visibility,pressure and conditions. 

This data is available as CSV file we are going to analyze this data set usng the panda dataframe.

find Out the diffrent analisys
Q) 1. Find All the unique "wind speed " Values in the data.
Q) 2. Find the number of time when the Weather if exactly Clear
Q) 3. Find the number of times when the "wind Speed was exactly 4 km/h"
Q. 4) Find out all the Null Values in the data
Q 5) Rename the Column Name "Weather" of the dataframe to "weather Condition"
Q. 6) What is the Mean "visibility"?
Q. 7) What is the Standard Deviation of Pressure in this data?
Q. 8) what is the Variance of "Relative Huminity" in this data?¶
Q. 9) Find all instances when 'Snow' was recorded.
Q. 10) Find all instances when 'Wind speed is above 24' and Visibilty is 25
Q. 11) What is the Mean value of each column against each 'Weather Condion'
Q. 12) What is the Minimum & Maximum values of each column against each "Weather Condition"
Q. 13) Show all the Record where Weather Condition is Fog.
Q. 14) Find all instances when "Weather is Clear " or "Visibility is above 40 ".
Q. 15 ) Find All instances When :
A. "Weather is clear " and "Relative Humidity is greater than 15"
or
B. "Visibility is above 40"